requirements.txt
# Core agent framework
langchain==0.3.0
langchain-anthropic==0.3.0
langchain-community==0.3.0
langgraph==0.2.0

# LLM
anthropic==0.40.0

# Data validation
pydantic==2.8.0
pydantic-settings==2.4.0

# Apache Spark
pyspark==3.5.0
delta-spark==3.2.0

# GCP clients
google-cloud-storage==2.18.0
google-cloud-dataproc==5.12.0
google-cloud-secret-manager==2.20.0
google-cloud-logging==3.11.0

# API framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0

# GitHub integration
PyGithub==2.3.0
gitpython==3.1.43

# Async / task queue
google-cloud-tasks==2.16.0

# Orchestration
apache-airflow==2.10.0

# Notebook support
jupyter==1.1.0
nbformat==5.10.0

# Data processing
pandas==2.2.0
pyarrow==17.0.0

# YAML / config
pyyaml==6.0.2
python-dotenv==1.0.1

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0
pytest-mock==3.14.0

# Code quality
black==24.8.0
pylint==3.2.0
mypy==1.11.0
isort==5.13.0

# Logging / observability
structlog==24.4.0
